extract SimpleSearchBar; use it in OverridesListScreen
