+++ inspektor/src/commonMain/kotlin/com/gyanoba/inspektor/ui/components/SimpleSearchBar.kt
@@ -1,0 +1,66 @@
+package com.gyanoba.inspektor.ui.components
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CornerSize
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.text.input.TextFieldState
+import androidx.compose.foundation.text.input.clearText
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Clear
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+internal fun SimpleSearchBar(
+    searchFieldState: TextFieldState,
+    placeholder: @Composable () -> Unit = { Text("Search") },
+    modifier: Modifier = Modifier,
+) {
+    val focusRequester = remember { FocusRequester() }
+    LaunchedEffect(Unit) {
+        focusRequester.requestFocus()
+    }
+    OutlinedTextField(
+        searchFieldState.text.toString(),
+        onValueChange = {
+            searchFieldState.edit {
+                replace(0, searchFieldState.text.length, it)
+            }
+        },
+        placeholder = placeholder,
+        keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Search),
+        shape = MaterialTheme.shapes.medium.copy(CornerSize(24.dp)),
+        trailingIcon = {
+            if (searchFieldState.text.isNotEmpty()) {
+                IconButton(
+                    onClick = { searchFieldState.clearText() },
+                    modifier = Modifier.size(24.dp),
+                ) {
+                    Icon(
+                        Icons.Rounded.Clear,
+                        tint = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.5f),
+                        contentDescription = "Clear",
+                        modifier = Modifier.size(16.dp)
+                    )
+                }
+            }
+        },
+        singleLine = true,
+        modifier = modifier.fillMaxWidth().padding(horizontal = 16.dp, vertical = 8.dp)
+            .focusRequester(focusRequester),
+    )
+}
