+++ inspektor/src/commonMain/kotlin/com/gyanoba/inspektor/ui/overriding/OverridesListViewModel.kt
@@ -1,5 +1,7 @@
 package com.gyanoba.inspektor.ui.overriding
 
+import androidx.compose.foundation.text.input.TextFieldState
+import androidx.compose.runtime.snapshotFlow
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.gyanoba.inspektor.data.HostMatcher
@@ -11,6 +13,7 @@ import com.gyanoba.inspektor.data.UrlRegexMatcher
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
@@ -22,32 +25,29 @@ internal class OverridesListViewModel(
 ) : ViewModel() {
     private val overrides = overrideRepository.updates
 
-    private val searchTerm = MutableStateFlow("")
+    val searchFieldState = TextFieldState()
+    private val searchTerm = snapshotFlow { searchFieldState.text }
 
     @OptIn(FlowPreview::class)
-    val visibleOverrides = overrides.debounce(500).map { overrides ->
-        if (searchTerm.value.isBlank()) {
-            return@map overrides
+    val visibleOverrides = overrides.debounce(500).combine(searchTerm) {  overrides, searchTerm ->
+        if (searchTerm.isBlank()) {
+            return@combine overrides
         }
 
         overrides.filter { override ->
-            (override.name?.contains(searchTerm.value, ignoreCase = true)
+            (override.name?.contains(searchTerm, ignoreCase = true)
                 ?: true) || override.matchers.any {
                 when (it) {
-                    is PathMatcher -> it.path.contains(searchTerm.value, ignoreCase = true)
-                    is HostMatcher -> it.host.contains(searchTerm.value, ignoreCase = true)
-                    is UrlMatcher -> it.url.contains(searchTerm.value, ignoreCase = true)
-                    is UrlRegexMatcher -> it.url.contains(searchTerm.value, ignoreCase = true)
+                    is PathMatcher -> it.path.contains(searchTerm, ignoreCase = true)
+                    is HostMatcher -> it.host.contains(searchTerm, ignoreCase = true)
+                    is UrlMatcher -> it.url.contains(searchTerm, ignoreCase = true)
+                    is UrlRegexMatcher -> it.url.contains(searchTerm, ignoreCase = true)
                 }
             }
         }
     }.stateIn(
         viewModelScope, SharingStarted.WhileSubscribed(4000), emptyList()
     )
-
-    fun search(query: String) {
-        searchTerm.value = query
-    }
 
     fun deleteOverride(override: Override) = viewModelScope.launch {
         overrideRepository.remove(override)
